Section 10 Resource based authorization for create, update, delete operation for Restaurant entity

+++ Restaurants.Application/Dishes/Commands/DeleteDishesForRestaurant/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,13 +1,16 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishesForRestaurant;
 public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRestaurantCommandHandler> logger,
 	IRestaurantsRepository restaurantsRepository,
-	IDishesRepository dishesRepository) : IRequestHandler<DeleteDishesForRestaurantCommand>
+	IDishesRepository dishesRepository,
+	IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteDishesForRestaurantCommand>
 {
 	public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
 	{
@@ -16,6 +19,12 @@ public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRest
 		var restaurant = await restaurantsRepository.GetAsync(request.RestaurantId);
 
 		if (restaurant == null) throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+
+
+		if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Update))
+		{
+			throw new ForbidException();
+		}
 
 		await dishesRepository.Delete(restaurant.Dishes);
 	}
