+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,11 +1,15 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
-public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger, IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
+public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger,
+	IRestaurantsRepository restaurantsRepository,
+	IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteRestaurantCommand>
 {
 	public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
 	{
@@ -16,6 +20,11 @@ public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandl
 		if (restaurant == null)
 		{
 			throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+		}
+
+		if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+		{
+			throw new ForbidException();
 		}
 
 		await restaurantsRepository.Delete(restaurant);
